fix(build): respect dependency order when building to prepare release
